WIP INT-3: FollowPagination setting and tests
Allow new pagination behavior to be turned off. This will allow users
to opt-out, if for some reason the new behavior doesn't align to
their needs. In the future, we may move to a finer-grained approach
to pagination, for now, it's on-and-automatic or off.

tests for pagination using a mocked API server