INT-3: follow Link headers for pagination (#76)
Applies to zones list, and zone get (for records). These methods follow Link
headers until things are complete.

Pagination following is on by default, but can be turned off, so users
can opt-out, if for some reason the new behavior doesn't align to
their needs. In the future, we may move to a finer-grained approach
to pagination, for now, it's on-and-automatic or off.

Tests for pagination use mocked API server