INBOX-344: make pagination match standard behavior a bit more (#77)
On HTTP error, the Do handler returns the response as well as the
error, and people were using that. In lieu of a breaking change,
we emulate this behavior as closely as we can in DoWithPagination.

DoWithPagination:
  * Return the response as well as the error, on HTTP Error
  * We now clobber the old responses as we traverse pages. On HTTP
    error, the response will match the error. Without error, the
    response will be the _last_ page of results.

Tried to exercise this in unit tests. Since it's a bit subtle, it
would be nice if a test broke if we accidentally change behavior.

Also a bunch of related improvements, mainly to tests:
* Refactored the little "mockAPI" stuff for more generic use
* Add some unit tests for the header parsing funcs
* Document that Do/DoWithPagination can return a nil or non-nil resp,
  depending on what kind of error happens.
* Add messages to some asserts
* unit tests in client_test for Do/DoWithPagination/getURI