INBOX-344: force HTTPS when following Link headers (#78)
When the Client endpoint is set to an HTTPS URL, rewrite any HTTP
Link URLs to HTTPS.

Note a small but potentially breaking change:

This commit changes the arguments to `ParseLinks`. This is a
very recently added function that should not need to be explicitly
called in normal SDK usage, but _is_ exported for convenience.